modified:   Assets/Fonts/fusion-pixel-12px SDF.asset modified:   Assets/Scripts/Gameplay/Entities/Enemy.cs modified:   Assets/Scripts/Gameplay/Entities/Entity.cs modified:   Assets/Scripts/Gameplay/Entities/EntitySave.cs modified:   Assets/Scripts/Gameplay/Entities/Inventory.cs modified:   Assets/Scripts/Gameplay/Entities/ItemBehaviour.cs modified:   Assets/Scripts/Gameplay/Entities/Player/Player.cs modified:   Assets/Scripts/Gameplay/Entities/Player/PlayerCenter.cs modified:   Assets/Scripts/Gameplay/Entities/Player/PlayerUI.cs modified:   Assets/Scripts/Gameplay/Entities/SkillManuscriptBehaviour.cs modified:   Assets/Scripts/Gameplay/WorldGeneration/IslandGeneration.cs modified:   Assets/Scripts/Mods/IDs/BlockID.cs modified:   Assets/Scripts/Mods/IDs/SkillID.cs modified:   Assets/Scripts/UI/NodeTree.cs

+++ Assets/Scripts/Gameplay/Entities/EntitySave.cs
@@ -49,7 +49,7 @@ namespace GameCore
         public Inventory inventory;// = new();
         public List<TaskStatusForSave> completedTasks = new();
         public List<SkillStatusForSave> unlockedSkills = new();
-        public int skillPoints;
+        public float skillPoints;
         public Vector2 respawnPoint;
 
         [NonSerialized] public byte[] skinHead;
